RAG-telegram-bot/requirements.txt
pymupdf>=1.24.0
sentence-transformers>=5.0.0
chromadb==1.5.0
rank-bm25>=0.2.2
groq>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
numpy>=1.26.0
httpx>=0.27.0,<1.0
torch>=2.0.0
python-telegram-bot>=21.0.0
